test: consolidate xarray fixtures in conftest.py (Story 4.5)
Add 3 temperature DataArray fixtures and promote 10 fixtures from
test_xarray_adapter.py to session-scoped conftest fixtures:

New fixtures added:
- sample_monthly_temp_da (1D, sinusoidal + noise)
- gridded_monthly_temp_3d (3D with latitude offset)
- dask_monthly_temp_3d (Dask-backed wrapper)

Promoted to conftest (session scope):
- coord_rich_1d_da, multi_coord_da (coordinate metadata)
- no_time_dim_da, non_monotonic_time_da, short_monthly_da (validation)
- sample_monthly_pet_offset_da (alignment tests)
- dask_multi_time_chunk, dask_monthly_pet_da (Dask tests)
- monthly_precip_with_nan, monthly_precip_heavy_nan (NaN handling)

Also clean up imports: add __future__ annotations, move pandas/xarray
to top-level imports.

Session-scoped fixtures reduce test overhead by constructing expensive
DataArrays once per session instead of per test function.

603 of 607 tests pass (4 pre-existing failures in test_xarray_equivalence.py).
