+++ tests/test_xarray_adapter.py
@@ -55,236 +55,6 @@ def _standard_normal(size: int | tuple[int, ...]) -> np.ndarray:
 
 
 # fixtures for test data specific to xarray_adapter tests
-@pytest.fixture
-def coord_rich_1d_da() -> xr.DataArray:
-    """Create 1D DataArray with rich coordinate metadata.
-
-    Features:
-    - time dimension coord with multiple attrs (axis, calendar, long_name, bounds, standard_name)
-    - month non-dimension auxiliary coord
-    - station_id scalar coord with attrs
-    """
-    # 36 months (3 years)
-    time = pd.date_range("2020-01-01", "2022-12-01", freq="MS")
-    rng = np.random.default_rng(42)
-    values = rng.gamma(shape=2.0, scale=50.0, size=len(time))
-
-    # auxiliary coordinate: month numbers
-    month = xr.DataArray(
-        [t.month for t in time],
-        dims=["time"],
-        attrs={"long_name": "month of year", "units": "1"},
-    )
-
-    # scalar coordinate with attrs
-    station_id = xr.DataArray(
-        "GHCN-12345",
-        attrs={"long_name": "station identifier", "cf_role": "timeseries_id"},
-    )
-
-    da = xr.DataArray(
-        values,
-        coords={
-            "time": time,
-            "month": month,
-            "station_id": station_id,
-        },
-        dims=["time"],
-        attrs={
-            "units": "mm",
-            "long_name": "Monthly Precipitation",
-        },
-        name="precip",
-    )
-
-    # add rich attrs to time coord
-    da.coords["time"].attrs = {
-        "axis": "T",
-        "calendar": "standard",
-        "long_name": "time",
-        "bounds": "time_bounds",
-        "standard_name": "time",
-    }
-
-    return da
-
-
-@pytest.fixture
-def multi_coord_da() -> xr.DataArray:
-    """Create 2D DataArray with multiple dimension coords and rich metadata.
-
-    Features:
-    - time and lat dimension coords, each with rich attrs
-    - month auxiliary coord (time-dependent)
-    - ensemble scalar coord with attrs
-    """
-    time = pd.date_range("2020-01-01", "2020-12-01", freq="MS")
-    lat = np.array([30.0, 35.0, 40.0, 45.0])
-    rng = np.random.default_rng(99)
-    values = rng.gamma(shape=2.0, scale=50.0, size=(len(time), len(lat)))
-
-    # auxiliary coordinate
-    month = xr.DataArray(
-        [t.month for t in time],
-        dims=["time"],
-        attrs={"long_name": "month of year"},
-    )
-
-    # scalar coordinate
-    ensemble = xr.DataArray(
-        "ens01",
-        attrs={"long_name": "ensemble member", "type": "hindcast"},
-    )
-
-    da = xr.DataArray(
-        values,
-        coords={
-            "time": time,
-            "lat": lat,
-            "month": month,
-            "ensemble": ensemble,
-        },
-        dims=["time", "lat"],
-        attrs={
-            "units": "mm",
-            "long_name": "Gridded Precipitation",
-        },
-        name="precip_grid",
-    )
-
-    # add rich attrs to dimension coords
-    da.coords["time"].attrs = {
-        "axis": "T",
-        "standard_name": "time",
-        "long_name": "time",
-    }
-    da.coords["lat"].attrs = {
-        "axis": "Y",
-        "standard_name": "latitude",
-        "long_name": "latitude",
-        "units": "degrees_north",
-    }
-
-    return da
-
-
-@pytest.fixture
-def no_time_dim_da() -> xr.DataArray:
-    """Create DataArray without a time dimension."""
-    # spatial coordinates only
-    x = np.arange(10)
-    lat = np.arange(5) * 10.0
-    rng = np.random.default_rng(42)
-    values = rng.gamma(shape=2.0, scale=50.0, size=(len(x), len(lat)))
-
-    return xr.DataArray(
-        values,
-        coords={"x": x, "lat": lat},
-        dims=["x", "lat"],
-        attrs={"units": "mm"},
-    )
-
-
-@pytest.fixture
-def non_monotonic_time_da() -> xr.DataArray:
-    """Create DataArray with non-monotonic (shuffled) time coordinate."""
-    time = pd.date_range("2020-01-01", "2020-12-01", freq="MS")
-    # shuffle the time coordinate
-    shuffled_time = time[[0, 2, 1, 3, 5, 4, 6, 8, 7, 9, 11, 10]]
-    rng = np.random.default_rng(42)
-    values = rng.gamma(shape=2.0, scale=50.0, size=len(time))
-
-    return xr.DataArray(
-        values,
-        coords={"time": shuffled_time},
-        dims=["time"],
-        attrs={"units": "mm"},
-    )
-
-
-@pytest.fixture
-def short_monthly_da() -> xr.DataArray:
-    """Create short DataArray with only 3 months."""
-    time = pd.date_range("2020-01-01", "2020-03-01", freq="MS")
-    rng = np.random.default_rng(42)
-    values = rng.gamma(shape=2.0, scale=50.0, size=len(time))
-
-    return xr.DataArray(
-        values,
-        coords={"time": time},
-        dims=["time"],
-        attrs={"units": "mm"},
-    )
-
-
-@pytest.fixture
-def sample_monthly_pet_offset_da() -> xr.DataArray:
-    """Create a 1D monthly PET DataArray with offset time range (1985-2024).
-
-    This fixture overlaps with sample_monthly_precip_da (1980-2019) only
-    during 1985-2019, useful for testing alignment behavior.
-    """
-    # 40 years offset by 5 years: 1985-2024
-    time = pd.date_range("1985-01-01", "2024-12-01", freq="MS")
-    # generate random PET values
-    rng = np.random.default_rng(101)
-    values = rng.gamma(shape=2.5, scale=60.0, size=len(time))
-
-    return xr.DataArray(
-        values,
-        coords={"time": time},
-        dims=["time"],
-        attrs={
-            "units": "mm",
-            "long_name": "Monthly Potential Evapotranspiration (Offset)",
-        },
-    )
-
-
-@pytest.fixture
-def dask_multi_time_chunk() -> xr.DataArray:
-    """Create 3-D Dask-backed DataArray with time split into multiple chunks (invalid).
-
-    This fixture intentionally violates the chunking constraint to test validation.
-    """
-    time = pd.date_range("1980-01-01", "2019-12-01", freq="MS")
-    lat = np.linspace(30.0, 50.0, 5)
-    lon = np.linspace(-120.0, -100.0, 6)
-    rng = np.random.default_rng(99)
-    values = rng.gamma(shape=2.0, scale=50.0, size=(len(time), len(lat), len(lon)))
-
-    da = xr.DataArray(
-        values,
-        coords={"time": time, "lat": lat, "lon": lon},
-        dims=["time", "lat", "lon"],
-        attrs={"units": "mm"},
-    )
-    # chunk time into multiple chunks (invalid for climate indices)
-    return da.chunk({"time": 120, "lat": 2, "lon": 3})
-
-
-@pytest.fixture
-def dask_monthly_pet_da() -> xr.DataArray:
-    """Create 1-D Dask-backed monthly PET DataArray matching dask_monthly_precip_1d.
-
-    For testing multi-input functions like SPEI with Dask arrays.
-    """
-    time = pd.date_range("1980-01-01", "2019-12-01", freq="MS")
-    rng = np.random.default_rng(100)
-    values = rng.gamma(shape=2.5, scale=60.0, size=len(time))
-
-    da = xr.DataArray(
-        values,
-        coords={"time": time},
-        dims=["time"],
-        attrs={
-            "units": "mm",
-            "long_name": "Monthly PET (Dask)",
-        },
-        name="pet_dask",
-    )
-    # chunk time as single chunk
-    return da.chunk({"time": -1})
 
 
 class TestXarrayAdapterNumpyPassthrough:
@@ -2670,50 +2440,6 @@ class TestSPEIIntegration:
 
 
 # NaN handling fixtures (Story 2.8)
-
-
-@pytest.fixture
-def monthly_precip_with_nan() -> xr.DataArray:
-    """Create 40-year monthly precipitation with ~10% NaN scattered throughout."""
-    time = pd.date_range("1980-01-01", "2019-12-01", freq="MS")
-    rng = np.random.default_rng(42)
-    values = rng.gamma(shape=2.0, scale=50.0, size=len(time))
-
-    # add scattered NaN (~10%)
-    nan_indices = rng.choice(len(time), size=int(len(time) * 0.1), replace=False)
-    values[nan_indices] = np.nan
-
-    return xr.DataArray(
-        values,
-        coords={"time": time},
-        dims=["time"],
-        attrs={
-            "units": "mm",
-            "long_name": "Monthly Precipitation with NaN",
-        },
-    )
-
-
-@pytest.fixture
-def monthly_precip_heavy_nan() -> xr.DataArray:
-    """Create 40-year monthly precipitation with ~80% NaN (insufficient data)."""
-    time = pd.date_range("1980-01-01", "2019-12-01", freq="MS")
-    rng = np.random.default_rng(123)
-    values = rng.gamma(shape=2.0, scale=50.0, size=len(time))
-
-    # add heavy NaN (~80%)
-    nan_indices = rng.choice(len(time), size=int(len(time) * 0.8), replace=False)
-    values[nan_indices] = np.nan
-
-    return xr.DataArray(
-        values,
-        coords={"time": time},
-        dims=["time"],
-        attrs={
-            "units": "mm",
-            "long_name": "Monthly Precipitation with Heavy NaN",
-        },
-    )
 
 
 # NaN handling test classes (Story 2.8)
